Put every Matrix's subclasses tests in the same file.
