+++ raytracing/tests/testsMatrix.py
@@ -1,5 +1,5 @@
 import unittest
-import envtest # modifies path
+import envtest  # modifies path
 
 from raytracing import *
 
@@ -474,167 +474,9 @@ class TestMatrix(unittest.TestCase):
         strRepresentation += "\nf = +inf (afocal)\n".format(-1.0)
         self.assertEqual(str(m).strip(), strRepresentation.strip())
 
-    def testDielectricInterface(self):
-        m = DielectricInterface(n1=1, n2=1.5, R=10)
-        self.assertEqual(m.determinant, 1 / 1.5)
-        self.assertEqual(m.frontVertex, 0)
-        self.assertEqual(m.backVertex, 0)
 
-    def testDielectricInterfaceConvergingSign(self):
-        # Positive R is convex for ray
-        m = DielectricInterface(n1=1, n2=1.5, R=10)
-        outRayDown = m*Ray(y=1,theta=0)
-        outRayUp = m*Ray(y=-1,theta=0)
 
-        # Ray is focussed to focal spot
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
 
-    def testDielectricInterfaceDivergingSign(self):
-        # Negative R is concave for ray
-        m = DielectricInterface(n1=1, n2=1.5, R=-10)
-        outRayUp = m*Ray(y=1,theta=0)
-        outRayDown = m*Ray(y=-1,theta=0)
-
-        # Ray is diverging
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
-
-    def testThickConvergingLens(self):
-        # Biconvex
-        m = ThickLens(n=1.55, R1=100, R2=-100, thickness=3)
-        outRayDown = m*Ray(y=1,theta=0)
-        outRayUp = m*Ray(y=-1,theta=0)
-
-        # Ray is focussed to focal spot
-        self.assertTrue(m.C < 0)
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
-
-    def testThickDivergingLens(self):
-        # Biconcave
-        m = ThickLens(n=1.55, R1=-100, R2=100, thickness=3)
-        outRayUp = m*Ray(y=1,theta=0)
-        outRayDown = m*Ray(y=-1,theta=0)
-
-        # Ray is diverging
-        self.assertTrue(m.C > 0)
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
-
-    def testThickConvergingLensEquivalence(self):
-        # Biconvex
-        m = ThickLens(n=1.55, R1=100, R2=-100, thickness=3)
-
-        mEquivalent = MatrixGroup()
-        mEquivalent.append(DielectricInterface(n1=1, n2=1.55, R=100))
-        mEquivalent.append(Space(d=3))
-        mEquivalent.append(DielectricInterface(n1=1.55, n2=1.0, R=-100))
-
-        self.assertAlmostEqual(m.A, mEquivalent.A,3)
-        self.assertAlmostEqual(m.B, mEquivalent.B,3)
-        self.assertAlmostEqual(m.C, mEquivalent.C,3)
-        self.assertAlmostEqual(m.D, mEquivalent.D,3)
-
-    def testThickConvergingLensFlip(self):
-        # Biconvex
-        m1 = ThickLens(n=1.55, R1=200, R2=-100, thickness=3)
-        m2 = ThickLens(n=1.55, R1=100, R2=-200, thickness=3)
-        m2.flipOrientation()
-
-        self.assertAlmostEqual(m1.determinant, 1,4)
-        self.assertAlmostEqual(m2.determinant, 1,4)
-        self.assertAlmostEqual(m1.A, m2.A,4)
-        self.assertAlmostEqual(m1.B, m2.B,4)
-        self.assertAlmostEqual(m1.C, m2.C,4)
-        self.assertAlmostEqual(m1.D, m2.D,4)
-
-    def testConvergingCurvedMirror(self):
-        # Concave should be positive?
-        m = CurvedMirror(R=-100)
-        outRayDown = m*Ray(y=1,theta=0)
-        outRayUp = m*Ray(y=-1,theta=0)
-
-        # Ray is focussed to focal spot
-        self.assertTrue(m.C < 0)
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
-
-    def testDivergingCurvedMirror(self):
-        m = CurvedMirror(R=100)
-        outRayUp = m*Ray(y=1,theta=0)
-        outRayDown = m*Ray(y=-1,theta=0)
-
-        # Ray is diverging
-        self.assertTrue(m.C > 0)
-        self.assertTrue(outRayDown.theta < 0)
-        self.assertTrue(outRayUp.theta > 0)
-
-    def testCurvedMirrorFlip(self):
-        # Biconvex
-        m1 = CurvedMirror(R=100)
-        m2 = CurvedMirror(R=-100)
-        m2.flipOrientation()
-
-        self.assertAlmostEqual(m1.determinant, 1,4)
-        self.assertAlmostEqual(m2.determinant, 1,4)
-        self.assertAlmostEqual(m1.A, m2.A,4)
-        self.assertAlmostEqual(m1.B, m2.B,4)
-        self.assertAlmostEqual(m1.C, m2.C,4)
-        self.assertAlmostEqual(m1.D, m2.D,4)
-
-    def testLensFocalLengths(self):
-        m = Lens(f=5)
-        self.assertEqual(m.effectiveFocalLengths(), (5, 5))
-        self.assertEqual(m.backFocalLength(), 5)
-        self.assertEqual(m.frontFocalLength(), 5)
-
-    def deactivated_testThickLensFocalLengths(self):
-        m = ThickLens(n=1.55, R1=100, R2=-100, thickness=3)
-
-        self.assertEqual(m.backFocalLength(), 5)
-        self.assertEqual(m.frontFocalLength(), 5)
-
-    def testOlympusLens(self):
-        self.assertIsNotNone(olympus.LUMPlanFL40X())
-        self.assertIsNotNone(olympus.XLUMPlanFLN20X())
-        self.assertIsNotNone(olympus.MVPlapo2XC())
-        self.assertIsNotNone(olympus.UMPLFN20XW())
-
-    def testThorlabsLenses(self):
-        l = thorlabs.ACN254_100_A()
-        l = thorlabs.ACN254_075_A()
-        l = thorlabs.ACN254_050_A()
-        l = thorlabs.ACN254_040_A()
-        l = thorlabs.AC254_030_A()
-        l = thorlabs.AC254_035_A()
-        l = thorlabs.AC254_045_A()
-        l = thorlabs.AC254_050_A()
-        l = thorlabs.AC254_060_A()
-        l = thorlabs.AC254_075_A()
-        l = thorlabs.AC254_080_A()
-        l = thorlabs.AC254_100_A()
-        l = thorlabs.AC254_125_A()
-        l = thorlabs.AC254_200_A()
-        l = thorlabs.AC254_250_A()
-        l = thorlabs.AC254_300_A()
-        l = thorlabs.AC254_400_A()
-        l = thorlabs.AC254_500_A()
-
-        l = thorlabs.AC508_075_B()
-        l = thorlabs.AC508_080_B()
-        l = thorlabs.AC508_100_B()
-        l = thorlabs.AC508_150_B()
-        l = thorlabs.AC508_200_B()
-        l = thorlabs.AC508_250_B()
-        l = thorlabs.AC508_300_B()
-        l = thorlabs.AC508_400_B()
-        l = thorlabs.AC508_500_B()
-        l = thorlabs.AC508_750_B()
-        l = thorlabs.AC508_1000_B()
-
-    def testEdmundLens(self):
-        l = eo.PN_33_921()
