+++ raytracing/tests/testsMatrixSubclasses.py
@@ -1,5 +1,5 @@
 import unittest
-import env  # modifies path
+import envtest  # modifies path
 
 from raytracing import *
 
@@ -51,14 +51,14 @@ class TestCurvedMirror(unittest.TestCase):
         cm = CurvedMirror(0.2, label="Test")
         self.assertEqual(cm.A, 1)
         self.assertEqual(cm.B, 0)
-        self.assertEqual(cm.C, -10)
+        self.assertEqual(cm.C, 10)
         self.assertEqual(cm.D, 1)
         self.assertEqual(cm.label, "Test")
 
         cm = CurvedMirror(-0.2, 9)
         self.assertEqual(cm.A, 1)
         self.assertEqual(cm.B, 0)
-        self.assertEqual(cm.C, 10)
+        self.assertEqual(cm.C, -10)
         self.assertEqual(cm.D, 1)
         self.assertEqual(cm.apertureDiameter, 9)
 
@@ -66,11 +66,70 @@ class TestCurvedMirror(unittest.TestCase):
         R = -4
         cm = CurvedMirror(R)
         z = 20
-        pointsInterest = [{'z': z - R / 2, 'label': "$F_f$"}, {'z': z + R / 2, 'label': "$F_b$"}]
+        pointsInterest = [{'z': z + R / 2, 'label': "$F_f$"}, {'z': z - R / 2, 'label': "$F_b$"}]
         self.assertListEqual(cm.pointsOfInterest(z), pointsInterest)
 
         cm.C = 0
         self.assertListEqual(cm.pointsOfInterest(z), [])
+
+    def testMatrix(self):
+        m = CurvedMirror(R=-10)
+        self.assertIsNotNone(m)
+
+    def testRayInMirror(self):
+        m1 = CurvedMirror(R=-10)
+        outRay1 = m1 * Ray(y=1, theta=0)
+        self.assertTrue(outRay1.theta < 0)
+
+        m2 = CurvedMirror(R=10)
+        outRay2 = m2 * Ray(y=1, theta=0)
+        self.assertTrue(outRay2.theta > 0)
+
+    def testRayInFlippedMirror(self):
+        m1 = CurvedMirror(R=-10)
+        outRay1 = m1 * Ray(y=1, theta=0)
+        m2 = CurvedMirror(R=10)
+        outRay2 = m2 * Ray(y=1, theta=0)
+
+        m3 = m2.flipOrientation()
+        outRay3 = m3 * Ray(y=1, theta=0)
+
+        self.assertEqual(outRay3.theta, outRay1.theta)
+        self.assertEqual(outRay3.theta, -outRay2.theta)
+
+    def testConvergingCurvedMirror(self):
+        # Concave should be positive?
+        m = CurvedMirror(R=-100)
+        outRayDown = m * Ray(y=1, theta=0)
+        outRayUp = m * Ray(y=-1, theta=0)
+
+        # Ray is focussed to focal spot
+        self.assertTrue(m.C < 0)
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
+    def testDivergingCurvedMirror(self):
+        m = CurvedMirror(R=100)
+        outRayUp = m * Ray(y=1, theta=0)
+        outRayDown = m * Ray(y=-1, theta=0)
+
+        # Ray is diverging
+        self.assertTrue(m.C > 0)
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
+    def testCurvedMirrorFlip(self):
+        # Biconvex
+        m1 = CurvedMirror(R=100)
+        m2 = CurvedMirror(R=-100)
+        m2.flipOrientation()
+
+        self.assertAlmostEqual(m1.determinant, 1, 4)
+        self.assertAlmostEqual(m2.determinant, 1, 4)
+        self.assertAlmostEqual(m1.A, m2.A, 4)
+        self.assertAlmostEqual(m1.B, m2.B, 4)
+        self.assertAlmostEqual(m1.C, m2.C, 4)
+        self.assertAlmostEqual(m1.D, m2.D, 4)
 
 
 class TestSpaceMatrix(unittest.TestCase):
@@ -208,6 +267,26 @@ class TestDielectricInterface(unittest.TestCase):
         self.assertEqual(di.C, -(1 - 1.33) / (-2 * 1))
         self.assertEqual(di.D, 1.33)
 
+    def testDielectricInterfaceConvergingSign(self):
+        # Positive R is convex for ray
+        m = DielectricInterface(n1=1, n2=1.5, R=10)
+        outRayDown = m * Ray(y=1, theta=0)
+        outRayUp = m * Ray(y=-1, theta=0)
+
+        # Ray is focussed to focal spot
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
+    def testDielectricInterfaceDivergingSign(self):
+        # Negative R is concave for ray
+        m = DielectricInterface(n1=1, n2=1.5, R=-10)
+        outRayUp = m * Ray(y=1, theta=0)
+        outRayDown = m * Ray(y=-1, theta=0)
+
+        # Ray is diverging
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
 
 class TestThickLens(unittest.TestCase):
 
@@ -248,6 +327,81 @@ class TestThickLens(unittest.TestCase):
         focusPos = p1 - f, p2 + f
         pointsInterest = [{'z': focusPos[0], 'label': "$F_f$"}, {'z': focusPos[1], 'label': "$F_b$"}]
         self.assertListEqual(tl.pointsOfInterest(z), pointsInterest)
+
+    def testMatrix(self):
+        m = ThickLens(R1=-10, R2=20, n=1.5, thickness=1)
+        self.assertIsNotNone(m)
+
+    def testRayConvergingThickLens(self):
+        m1 = ThickLens(R1=10, R2=20, n=1.5, thickness=1)
+        outRay1 = m1 * Ray(y=1, theta=0)
+        self.assertTrue(outRay1.theta < 0)
+
+    def testRayDivergingThickLens(self):
+        m2 = ThickLens(R1=-10, R2=-20, n=1.5, thickness=1)
+        outRay2 = m2 * Ray(y=1, theta=0)
+        self.assertTrue(outRay2.theta > 0)
+
+    def testRayInFlippedThickLens(self):
+        m1 = CurvedMirror(R=-10)
+        outRay1 = m1 * Ray(y=1, theta=0)
+        m2 = CurvedMirror(R=10)
+        outRay2 = m2 * Ray(y=1, theta=0)
+
+        m3 = m2.flipOrientation()
+        outRay3 = m3 * Ray(y=1, theta=0)
+
+        self.assertEqual(outRay3.theta, outRay1.theta)
+        self.assertEqual(outRay3.theta, -outRay2.theta)
+
+    def testThickConvergingLens(self):
+        # Biconvex
+        m = ThickLens(n=1.55, R1=100, R2=-100, thickness=3)
+        outRayDown = m * Ray(y=1, theta=0)
+        outRayUp = m * Ray(y=-1, theta=0)
+
+        # Ray is focussed to focal spot
+        self.assertTrue(m.C < 0)
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
+    def testThickDivergingLens(self):
+        # Biconcave
+        m = ThickLens(n=1.55, R1=-100, R2=100, thickness=3)
+        outRayUp = m * Ray(y=1, theta=0)
+        outRayDown = m * Ray(y=-1, theta=0)
+
+        # Ray is diverging
+        self.assertTrue(m.C > 0)
+        self.assertTrue(outRayDown.theta < 0)
+        self.assertTrue(outRayUp.theta > 0)
+
+    def testThickConvergingLensEquivalence(self):
+        # Biconvex
+        m = ThickLens(n=1.55, R1=100, R2=-100, thickness=3)
+
+        mEquivalent = MatrixGroup()
+        mEquivalent.append(DielectricInterface(n1=1, n2=1.55, R=100))
+        mEquivalent.append(Space(d=3))
+        mEquivalent.append(DielectricInterface(n1=1.55, n2=1.0, R=-100))
+
+        self.assertAlmostEqual(m.A, mEquivalent.A, 3)
+        self.assertAlmostEqual(m.B, mEquivalent.B, 3)
+        self.assertAlmostEqual(m.C, mEquivalent.C, 3)
+        self.assertAlmostEqual(m.D, mEquivalent.D, 3)
+
+    def testThickConvergingLensFlip(self):
+        # Biconvex
+        m1 = ThickLens(n=1.55, R1=200, R2=-100, thickness=3)
+        m2 = ThickLens(n=1.55, R1=100, R2=-200, thickness=3)
+        m2.flipOrientation()
+
+        self.assertAlmostEqual(m1.determinant, 1, 4)
+        self.assertAlmostEqual(m2.determinant, 1, 4)
+        self.assertAlmostEqual(m1.A, m2.A, 4)
+        self.assertAlmostEqual(m1.B, m2.B, 4)
+        self.assertAlmostEqual(m1.C, m2.C, 4)
+        self.assertAlmostEqual(m1.D, m2.D, 4)
 
 
 class TestDielectricSlab(unittest.TestCase):
